+++ packages/contracts/src/gateway/event/request/user-events.request.ts
@@ -1,0 +1,7 @@
+import type { PaginationRequest } from '../../../common/pagination.js';
+
+export interface GetUserEventsRequest extends Partial<PaginationRequest> {}
+
+export interface GetUserParticipationsRequest extends Partial<PaginationRequest> {}
+
+export interface GetUserWishesRequest extends Partial<PaginationRequest> {}